% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graficas.R
\name{mod_graficas_ui}
\alias{mod_graficas_ui}
\alias{mod_graficas_server}
\title{mod_graficas_ui and mod_graficas_server}
\usage{
mod_graficas_ui(id)

mod_graficas_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
