% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicadores.R
\name{mod_indicadores_ui}
\alias{mod_indicadores_ui}
\alias{mod_indicadores_server}
\title{mod_indicadores_ui and mod_indicadores_server}
\usage{
mod_indicadores_ui(id)

mod_indicadores_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
