% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa.R
\name{mod_mapa_ui}
\alias{mod_mapa_ui}
\alias{mod_mapa_server}
\title{mod_mapa_ui and mod_mapa_server}
\usage{
mod_mapa_ui(id)

mod_mapa_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
