% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tablas.R
\name{mod_tablas_ui}
\alias{mod_tablas_ui}
\alias{mod_tablas_server}
\title{mod_tablas_ui and mod_tablas_server}
\usage{
mod_tablas_ui(id)

mod_tablas_server(input, output, session, react)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
